feat: extract transcoding into dedicated worker container (#213)
- Create shared/ package with common models, database, config, storage,
  probe, and image utilities used by both CMS and worker
- Create worker/ package with FFmpeg transcoder, LISTEN/NOTIFY support,
  and Azure Storage Queue mode for scale-to-zero Container Apps Jobs
- Refactor CMS to use shared/ via re-export shims for full backward
  compatibility (all existing imports continue to work)
- Remove transcoder_loop from CMS startup (worker handles transcoding)
- Add NOTIFY signaling in CMS routers after variant enqueue/reset
- Add Dockerfile.worker for dedicated transcode container
- Update docker-compose.yml with worker service sharing asset volume
- Update all test imports to reference worker.transcoder for moved
  functions (_build_ffmpeg_args_safe, _transcode_one)
- Update cancel tests to reflect CMS no-op cancel (worker manages
  its own ffmpeg processes)

818 tests passing (1 pre-existing scheduler FK test excluded)
