+++ cms/database.py
@@ -1,39 +1,11 @@
-"""Database engine and session management."""
+"""Database engine and session management.
 
-from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
-from sqlalchemy.orm import DeclarativeBase
+Re-exports shared database primitives and adds CMS-only migrations.
+"""
 
-from cms.config import Settings
-
-
-class Base(DeclarativeBase):
-    pass
-
-
-_engine = None
-_session_factory = None
-
-
-def init_db(settings: Settings):
-    global _engine, _session_factory
-    _engine = create_async_engine(settings.database_url, echo=False)
-    _session_factory = async_sessionmaker(_engine, expire_on_commit=False)
-
-
-async def dispose_db():
-    global _engine
-    if _engine:
-        await _engine.dispose()
-
-
-async def get_db() -> AsyncSession:
-    async with _session_factory() as session:
-        yield session
-
-
-async def create_tables():
-    async with _engine.begin() as conn:
-        await conn.run_sync(Base.metadata.create_all)
+from shared.database import Base, init_db, get_db, dispose_db, create_tables  # noqa: F401
+from shared.database import get_engine, get_session_factory  # noqa: F401
+from shared import database as _shared_db
 
 
 async def run_migrations():
@@ -46,10 +18,10 @@ async def run_migrations():
 
     # Create any brand-new tables first so FK references in ALTER statements
     # can resolve (e.g. api_keys.user_id → users.id on first RBAC deploy).
-    async with _engine.begin() as conn:
+    async with _shared_db._engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
 
-    async with _engine.begin() as conn:
+    async with _shared_db._engine.begin() as conn:
         # Helper to check if a column exists in a table
         def _has_column(connection, table_name, column_name):
             insp = sa_inspect(connection)
@@ -280,5 +252,5 @@ async def run_migrations():
             ))
 
     # Run create_all again in case migrations added models with new relationships
-    async with _engine.begin() as conn:
+    async with _shared_db._engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
