+++ cms/models/group_asset.py
@@ -1,35 +1,5 @@
-"""GroupAsset junction model — tracks asset-to-group associations."""
+"""GroupAsset model — re-exported from shared package for backward compatibility."""
 
-import uuid
-from datetime import datetime, timezone
+from shared.models.group_asset import GroupAsset
 
-from sqlalchemy import DateTime, ForeignKey, UniqueConstraint
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-
-from cms.database import Base
-
-
-class GroupAsset(Base):
-    """Links an asset to a device group."""
-
-    __tablename__ = "group_assets"
-    __table_args__ = (
-        UniqueConstraint("asset_id", "group_id", name="uq_group_asset"),
-    )
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    asset_id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), ForeignKey("assets.id", ondelete="CASCADE"), nullable=False
-    )
-    group_id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), ForeignKey("device_groups.id", ondelete="CASCADE"), nullable=False
-    )
-    shared_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-
-    asset: Mapped["Asset"] = relationship(back_populates="group_asset_links")
-    group: Mapped["DeviceGroup"] = relationship()
+__all__ = ["GroupAsset"]
