+++ requirements-shared.txt
@@ -1,0 +1,9 @@
+# Shared package dependencies (used by both CMS and worker)
+pydantic>=2.10,<3.0
+pydantic-settings>=2.7,<3.0
+sqlalchemy[asyncio]>=2.0,<3.0
+asyncpg>=0.30,<1.0
+
+# Azure Blob Storage (only needed when AGORA_CMS_STORAGE_BACKEND=azure)
+azure-storage-blob>=12.20,<13.0
+aiohttp>=3.9,<4.0
