+++ shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared package — models, config, database, and services used by both CMS and worker."""
