+++ shared/database.py
@@ -1,0 +1,46 @@
+"""Database engine and session management (shared between CMS and worker)."""
+
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
+from sqlalchemy.orm import DeclarativeBase
+
+from shared.config import SharedSettings
+
+
+class Base(DeclarativeBase):
+    pass
+
+
+_engine = None
+_session_factory = None
+
+
+def init_db(settings: SharedSettings):
+    global _engine, _session_factory
+    _engine = create_async_engine(settings.database_url, echo=False)
+    _session_factory = async_sessionmaker(_engine, expire_on_commit=False)
+
+
+def get_engine():
+    """Return the current async engine (for raw connection access)."""
+    return _engine
+
+
+def get_session_factory():
+    """Return the current session factory."""
+    return _session_factory
+
+
+async def dispose_db():
+    global _engine
+    if _engine:
+        await _engine.dispose()
+
+
+async def get_db() -> AsyncSession:
+    async with _session_factory() as session:
+        yield session
+
+
+async def create_tables():
+    async with _engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
