+++ shared/models/__init__.py
@@ -1,0 +1,3 @@
+from shared.models.asset import Asset, AssetType, AssetVariant, DeviceAsset, VariantStatus  # noqa: F401
+from shared.models.device_profile import DeviceProfile  # noqa: F401
+from shared.models.group_asset import GroupAsset  # noqa: F401
