+++ worker/__init__.py
@@ -1,0 +1,1 @@
+"""Transcode worker package — dedicated FFmpeg video/image transcoding."""
